Reorder tracy_lz4.cpp vs TracyProfiler.cpp in TracyClient.cpp.
This fixes deprecation warning in tracy_lz4.hpp, which was previously
present due to TracyProfiler.cpp including tracy_lz4.hpp before an
appropriate deprecation restraining macro was defined in tracy_lz4.cpp.
Note that this issue was only present if TracyClient.cpp was used to
include the profiler in a project. Including the profiler as a
collection of separate source files worked correctly, as the deprecated
function is only used by tracy_lz4.cpp.