Added install_vcpkg_dependencies.bat to set up dependencies
This script will build both vcpkg and the dependencies needed by Tracy. It puts everything in the vcpkg/vcpkg directory, and changes no other state on the machine. It is perfectly safe to erase the vcpkg/vcpkg directory and re-run the script to build the dependencies again, although that should only be needed once.

To add new vcpkg dependencies, just modify the 'vcpkg install ...' line in the .bat file.