Updated Visual Studio project files to use vcpkg directory for dependencies
This modifies all of the include and lib paths to point to vcpkg/vcpkg/installed/x64-windows-static/include and lib.

With these changes, all executables in Tracy build out of the box in all configurations (assuming you have run the install_vcpkg_depencies.bat script first).

Perhaps it would be better to use a single Visual Studio .props file that all of the .vcxproj files point to so that the include and lib paths are set in a single place, but for now the paths are set separately in each .vcxproj.