This commit makes the D3D11 code consistent with the D3D12 code. It now properly supports TRACY_CALLSTACK as well as the Tracy*Transient macros.
The order of the macros now has the same order as the D3D12 file.

Added support for TracyD3D11ContextName().

I removed TRACY_D3D11_NO_SINGLE_THREAD, because I'm not
sure what that is supposed to be used for. It was set
up in an upstream fork of Tracy.