Fix horizontal scroll of main window.
Horizontal mouse scroll (or touchpad action) should map to navigation within
the trace. While this was true in most cases, there was an edge case, when the
top buttons row couldn't be entirely displayed, e.g. due to the window being
resized too small, or due to set zoom level. This change disables scroll
action on the main window, which fixes the issue.