Suppress variable shadowing warnings
This disables the warnings for MSVC, GCC and Clang in the ZoneScopedXX
macros.

The warnings produced are both a false positive since they didn't find a
bug *and* they don't happen in user written code, so the user couldn't
even do much about it. The previous workaround of using ZoneNamedXXX is
a poor solution since the Zone(Text|Name|etc.) macros all rely on the
`___tracy_scoped_zone` name.