Eliminate the function GetZoneCtx
The function GetZoneCtx isn't needed. This patch introduces a small
structure to help transport the context around to avoid the expensive
search. The GPU timeline function take the context as a parameter now
instead of the thread ID, since the supplied thread ID is derived from
the context in every invocation.