extra: Add a reference OpenCL layer implementation
This adds a proof-of-concept of an OpenCL layer [0], a feature provided
by OpenCL-ICD-Loader, that adds the capability to trace OpenCL calls
without direct modifications to application code.

The current version only wraps around the

- clEnqueueNDRangeKernel
- clFinish

functions and is limited to 32 combinations of cl_context and
cl_device_id for getting their appropriate TracyCLCtx.

[0] https://github.com/KhronosGroup/OpenCL-ICD-Loader#about-layers