extra: Add a reference OpenCL layer implementation
This adds a proof-of-concept of an OpenCL layer [0], a feature provided
by OpenCL-ICD-Loader, that adds the capability to trace OpenCL calls
without direct modifications to application code.

The current version wraps around the

- Buffer Objects APIs,
- Memory Objects APIs,
- Executing Kernels APIs, and
- Program Objects API.

The current version supports only 32 combinations of cl_context and
cl_device_id when retrieving the appropriate TracyCLCtx.

[0] https://github.com/KhronosGroup/OpenCL-ICD-Loader#about-layers